internal/shellexec: add NoopCommand test helper function
